requirements.txt
asgiref==3.5.2
bottle==0.12.19
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.0.12
conan==1.48.0
cryptography==37.0.2
defusedxml==0.7.1
distro==1.6.0
dj-database-url==0.5.0
Django==4.0.5
django-allauth==0.51.0
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-on-heroku==1.1.2
django-widget-tweaks==1.4.12
Jinja2==3.1.2
node-semver==0.6.1
patch-ng==1.17.4
Pillow==9.1.1
pluginbase==1.0.1
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.28.0
requests-oauthlib==1.3.1
tqdm==4.64.0
urllib3==1.26.9
whitenoise==6.2.0
